+++ bundlewrap/cmdline/items.py
@@ -24,7 +24,7 @@ def write_preview(file_item, base_path):
 
 
 def bw_items(repo, args):
-    node = get_node(repo, args['node'], adhoc_nodes=args['adhoc_nodes'])
+    node = get_node(repo, args['node'])
     if args['preview'] and not args['item']:
         io.stderr(_("{x} no ITEM given for preview").format(x=red("!!!")))
         exit(1)