+++ bundlewrap/repo.py
@@ -294,14 +294,6 @@ class Repository:
 
         open(join(bundle_dir, FILENAME_BUNDLE), 'a').close()
 
-    def create_node(self, node_name):
-        """
-        Creates an adhoc node with the given name.
-        """
-        node = Node(node_name)
-        self.add_node(node)
-        return node
-
     def get_all_attrs_from_file(self, path, base_env=None):
         """
         Reads all 'attributes' (if it were a module) from a source file.